change inc_alloc type to int64_t

+++ src/common/fdir_types.h
@@ -132,7 +132,7 @@ typedef union {
 typedef struct fdir_set_dentry_size_info {
     uint64_t inode;
     int64_t file_size;
-    int inc_alloc;
+    int64_t inc_alloc;
     bool force;
     int flags;
 } FDIRSetDEntrySizeInfo;
